Add support for time zone offsets
FlySight normally names files and sets modification times using UTC. This
changeset adds a TZ_Offset configuration value, allowing the user to work
in local time.

The logged data retains the same format, e.g. "2013-09-04T02:02:39.50Z",
regardless of TZ_Offset, but this way a day's worth of skydives will be
stored in the same directory and with filesystem timestamps suitable for
correlating with video files or other captured data.