Move version into FLYSIGHT_VERSION define
This is #defined from the Makefile by passing in a -D compiler option.
The version is generated automatically using `git describe`, the output
of which uniquely identifies the commit. Additionally, the output can
be made more readable (i.e. humanized) by using tags.

Changing the firmware version no longer involves updating any files; it
merely requires committing and building. As a result, Version.h must
always be considered "dirty" by the build system, and any referencing
code must be rebuilt. Recompilation could be reduced by defining
FlysightVersion as an extern char[] PROGMEM, but this is slightly less
convenient for the calling code, so I decided to go this route.

This can be made to work on Windows using batch files, but I do not have
a suitable development environment at this time.